more changes

+++ components/ui/Components/customList.jsx
@@ -6,12 +6,7 @@ import { capitalizeFirstLetter } from "../../../src/utils/functiones/functions";
 import { useFocusEffect, useNavigation } from "@react-navigation/native";
 import { ModalComponente } from "./customModal";
 import { TouchableOpacity } from "react-native-gesture-handler";
-import { refreshControl } from "../../../src/utils/functiones/refresh";
-import { NotRegistration } from "../../share/noRegistration";
-import { DeleteConfirmation } from "../../share/deletePress";
-import { InfoDocente } from "../(DIRECTOR)/docentes/components/inforDocente";
-import { ViewDocente } from "../(DIRECTOR)/docentes/components/viewDocente";
-import CustomTouchableOpacity from "./customTouchableOpacity";
+
 export const ListItemComponent = ({
   getDataAll,
   getDataOne,
@@ -24,7 +19,7 @@ export const ListItemComponent = ({
   const [items, setItems] = useState([]);
   const [modalVisible, setModalVisible] = useState(false);
   const [selectedItem, setSelectedItem] = useState(null);
-  console.log(selectedItem, "setSelectedItem");
+  console.log(selectedItem, 'setSelectedItem')
   const [refreshing, setRefreshing] = useState(false);
 
   const fetchItems = useCallback(async () => {
@@ -43,7 +38,9 @@ export const ListItemComponent = ({
       setModalVisible(true);
       const res = await getDataOne(cedula);
       console.log(res, "response res de handleInfoPress");
-      const itemselected = res.find((value) => value.cedula === cedula);
+      const itemselected = res.find(
+        (value) => value.cedula === cedula
+      );
       console.log(itemselected, "item selected");
       if (itemselected) {
         setSelectedItem(itemselected);
@@ -62,13 +59,29 @@ export const ListItemComponent = ({
   };
 
   const handleDeletePress = (itemId) => {
-    DeleteConfirmation({
-      nameDelete: modalTitle,
-      onPress: async () => {
-        await deleteData(itemId);
-        setItems(items.filter((item) => item.cedula !== itemId));
-      },
-    });
+    Alert.alert(
+      `Eliminar ${modalTitle}`,
+      `¿Estás seguro de que deseas eliminar este ${modalTitle.toLowerCase()}?`,
+      [
+        {
+          text: "Cancelar",
+          style: "cancel",
+        },
+        {
+          text: "Eliminar",
+          style: "destructive",
+          onPress: async () => {
+            try {
+              await deleteData(itemId);
+              setItems(items.filter((item) => item.cedula !== itemId));
+              Alert.alert(`${modalTitle} eliminado con éxito`);
+            } catch (error) {
+              Alert.alert(`Error al eliminar el ${modalTitle.toLowerCase()}`);
+            }
+          },
+        },
+      ]
+    );
   };
 
   const onRefresh = useCallback(async () => {
@@ -78,9 +91,20 @@ export const ListItemComponent = ({
   }, [fetchItems]);
 
   return (
-    <ScrollView refreshControl={refreshControl(refreshing, onRefresh)}>
+    <ScrollView
+      refreshControl={
+        <RefreshControl
+          refreshing={refreshing}
+          colors={["#78e08f"]}
+          onRefresh={onRefresh}
+          progressBackgroundColor="#1371C3"
+        />
+      }
+    >
       {items.length === 0 ? (
-        <NotRegistration />
+        <Text style={{ textAlign: "center", marginTop: 20 }}>
+          Ningún registro
+        </Text>
       ) : (
         items.map((item, index) => (
           <ListItem.Swipeable
@@ -101,7 +125,7 @@ export const ListItemComponent = ({
                 title="Delete"
                 onPress={() => {
                   reset();
-                  handleDeletePress(item.cedula); // aca debe ir DeleteConfirmation
+                  handleDeletePress(item.cedula);
                 }}
                 icon={{ name: "delete", color: "white" }}
                 buttonStyle={{ minHeight: "100%", backgroundColor: "red" }}
@@ -111,14 +135,23 @@ export const ListItemComponent = ({
             <Icon name={itemIcon} size={25} color="black" />
             <ListItem.Content>
               <ListItem.Title>
-                <CustomTouchableOpacity
-                  navigateToFormScreen={navigateToFormScreen}
-                  screenName="FormScreen"
-                  paramKey="cedula"
-                  paramValue={item.cedula}
+                <TouchableOpacity
+                  className="flex-row"
+                  onPress={() =>
+                    navigateToFormScreen
+                      ? navigateToFormScreen(navigation, item.cedula)
+                      : navigation.navigate("FormScreen", {
+                          cedula: item.cedula,
+                        })
+                  }
                 >
-                  <ViewDocente item={item} />
-                </CustomTouchableOpacity>
+                  <Text className="font-extrabold text-lg">
+                    {capitalizeFirstLetter(item.nombre)}{" "}
+                  </Text>
+                  <Text className="font-extrabold text-lg">
+                    {capitalizeFirstLetter(item.apellido)}
+                  </Text>
+                </TouchableOpacity>
               </ListItem.Title>
             </ListItem.Content>
             <ListItem.Chevron />
@@ -132,11 +165,205 @@ export const ListItemComponent = ({
         handleCloseModal={handleCloseModal}
       >
         {selectedItem ? (
-          <InfoDocente selectedItem={selectedItem} />
+          <>
+            <Text style={{ fontSize: 22, fontWeight: "bold" }}>
+              {capitalizeFirstLetter(selectedItem.nombre)}{" "}
+              {capitalizeFirstLetter(selectedItem.apellido)}
+            </Text>
+            <Text style={{ fontSize: 16, fontWeight: "bold" }}>
+              Cédula:{" "}
+              <Text className="font-normal">{selectedItem.cedula}</Text>
+            </Text>
+            <Text style={{ fontSize: 16, fontWeight: "bold" }}>
+              Correo:{" "}
+              <Text className="font-normal">{selectedItem.correo}</Text>
+            </Text>
+          </>
         ) : (
-          <NotRegistration />
+          <Text>No hay datos disponibles</Text>
         )}
       </ModalComponente>
     </ScrollView>
   );
 };
+
+
+// import { Alert, ScrollView, Text } from "react-native";
+// import { ListItem, Button } from "@rneui/themed";
+// import Icon from "@expo/vector-icons/MaterialCommunityIcons";
+// import { useCallback, useState } from "react";
+// import { useFocusEffect ,useNavigation} from "@react-navigation/native";
+// import { ModalComponente } from "./customModal";
+// import { refreshControl } from "../../../src/utils/functiones/refresh";
+// import { NotRegistration } from "../../share/noRegistration";
+// // import { DeleteConfirmation } from "../../share/deletePress";
+// import { InfoDocente } from "../(DIRECTOR)/docentes/components/inforDocente";
+// import { ViewDocente } from "../(DIRECTOR)/docentes/components/viewDocente";
+// import CustomTouchableOpacity from "./customTouchableOpacity";
+// import { TouchableOpacity } from "react-native-gesture-handler";
+// import { capitalizeFirstLetter } from "../../../src/utils/functiones/functions";
+// // import useToastMessage from "../../share/ToasNotification";
+// export const ListItemComponent = ({
+//   getDataAll,
+//   getDataOne,
+//   deleteData,
+//   navigateToFormScreen,
+//   itemIcon = "account",
+//   modalTitle = "Info",
+// }) => {
+//   // const { showToast, APP_STATUS, STATUS_MESSAGES } = useToastMessage();
+//   const navigation = useNavigation();
+//   const [items, setItems] = useState([]);
+//   const [modalVisible, setModalVisible] = useState(false);
+//   const [selectedItem, setSelectedItem] = useState(null);
+//   console.log(selectedItem, "setSelectedItem");
+//   const [refreshing, setRefreshing] = useState(false);
+
+//   const fetchItems = useCallback(async () => {
+//     const res = await getDataAll();
+//     setItems(res);
+//   }, [getDataAll]);
+
+//   useFocusEffect(
+//     useCallback(() => {
+//       fetchItems();
+//     }, [fetchItems])
+//   );
+
+//   const onRefresh = useCallback(async () => {
+//     setRefreshing(true);
+//     await fetchItems();
+//     setRefreshing(false);
+//   }, [fetchItems]);
+  
+//   const handleInfoPress = async (cedula) => {
+//     try {
+//       setModalVisible(true);
+//       const res = await getDataOne(cedula);
+//       console.log(res, "response res de handleInfoPress");
+//       const itemselected = res.find((value) => value.cedula === cedula);
+//       console.log(itemselected, "item selected");
+//       if (itemselected) {
+//         setSelectedItem(itemselected);
+//       } else {
+//         console.error("Docente no encontrado");
+//         setSelectedItem(null);
+//       }
+//     } catch (error) {
+//       throw new Error("Error fetching item:", error);
+//     }
+//   };
+
+//   const handleCloseModal = () => {
+//     setModalVisible(false);
+//     setSelectedItem(null);
+//   };
+
+//   const handleDeletePress = (itemId) => {
+//     Alert.alert(
+//       `Eliminar ${modalTitle}`,
+//       `¿Estás seguro de que deseas eliminar este ${modalTitle.toLowerCase()}?`,
+//       [
+//         {
+//           text: "Cancelar",
+//           style: "cancel",
+//         },
+//         {
+//           text: "Eliminar",
+//           style: "destructive",
+//           onPress: async () => {
+//             try {
+//               await deleteData(itemId);
+//               setItems(items.filter((item) => item.cedula !== itemId));
+//               Alert.alert(`${modalTitle} eliminado con éxito`);
+//             } catch (error) {
+//               Alert.alert(`Error al eliminar el ${modalTitle.toLowerCase()}`);
+//             }
+//           },
+//         },
+//       ]
+//     );
+//   };
+
+
+//   return (
+//     <ScrollView refreshControl={refreshControl(refreshing, onRefresh)}>
+//       {items.length === 0 ? (
+//         <NotRegistration />
+//       ) : (
+//         items.map((item, index) => (
+//           <ListItem.Swipeable
+//             key={item.id || index}
+//             leftContent={(reset) => (
+//               <Button
+//                 title="Info"
+//                 onPress={async () => {
+//                   reset();
+//                   await handleInfoPress(item.cedula);
+//                 }}
+//                 icon={{ name: "info", color: "white" }}
+//                 buttonStyle={{ minHeight: "100%" }}
+//               />
+//             )}
+//             rightContent={(reset) => (
+//               <Button
+//                 title="Delete"
+//                 onPress={() => {
+//                   reset();
+//                   handleDeletePress(item.cedula)}}
+//                 icon={{ name: "delete", color: "white" }}
+//                 buttonStyle={{ minHeight: "100%", backgroundColor: "red" }}
+//               />
+//             )}
+//           >
+//             <Icon name={itemIcon} size={25} color="black" />
+//             <ListItem.Content>
+//               <ListItem.Title>
+//                 {/* <CustomTouchableOpacity
+//                   navigateToFormScreen={navigateToFormScreen}
+//                   screenName="FormScreen"
+//                   paramKey="cedula"
+//                   paramValue={item.cedula}
+//                 >
+//                   <ViewDocente item={item} />
+//                 </CustomTouchableOpacity> */}
+//                 <TouchableOpacity
+//                   className="flex-row"
+//                   onPress={() =>
+//                     navigateToFormScreen
+//                       ? navigateToFormScreen(navigation, item.cedula)
+//                       : navigation.navigate("FormScreen", {
+//                           cedula: item.cedula,
+//                         })
+//                   }
+//                 >
+//                   <Text className="font-extrabold text-lg">
+//                     {capitalizeFirstLetter(item.nombre)}{" "}
+//                   </Text>
+//                   <Text className="font-extrabold text-lg">
+//                     {capitalizeFirstLetter(item.apellido)}
+//                   </Text>
+                  
+//                 </TouchableOpacity>
+//               </ListItem.Title>
+//             </ListItem.Content>
+//             <ListItem.Chevron />
+//           </ListItem.Swipeable>
+//         ))
+//       )}
+//       <ModalComponente
+//         transparent={true}
+//         animationType={"slider"}
+//         modalVisible={modalVisible}
+//         handleCloseModal={handleCloseModal}
+//       >
+//         {selectedItem ? (
+//           <InfoDocente selectedItem={selectedItem} />
+//         ) : (
+//           <NotRegistration />
+//         )}
+//       </ModalComponente>
+//     </ScrollView>
+//   );
+// };
+
